Internal: UI refactoring [ED-14060]
